Add Lotus Launch Specifications to the documents Specifications
As per title. These currently only include specifications for items
which diverge from the BCH/BCHA blockchains. However, the structure
of the folder is to be topical-based rather than activation time based.
Each future commit should be a single commit which includes all the
changes and activation times of new features.

This enables easy updating of the specifications, as well as an easy
to use reference -- rather than needing to look through activation-time
organized specifications.