Remove calculator (#82)
This isn't something that should drive people to mine. We already saw
that when there price was too high merc miners came in and drove price
back down. That generally means the stability mechanism is working as
intended, but this calculator is probably not accurate, nor is it
something that we should promote on the main site. If Lotus become
profitable for miners there will be other websites that pop up.